Sortirana lista like obavestenja prema datumu

+++ BusinesLogicLayer/Implementation/LikeNotificationService.cs
@@ -1,6 +1,8 @@
 ﻿using BusinesLogicLayer.Interfaces;
 using DataAccessLayer.UnitOfWork;
 using Domain;
+using Dto;
+using Mapper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,10 +14,12 @@ namespace BusinesLogicLayer.Implementation
     public class LikeNotificationService : ILikeNotificationService
     {
         private readonly IUnitOfWork unit;
+        private LikeNotificationMapper mapper;
 
         public LikeNotificationService(UserContext context)
         {
             this.unit = new DataAccessLayer.UnitOfWork.UnitOfWork(context);
+            mapper = new LikeNotificationMapper();
         }
 
         public bool Create(Notification entity)
@@ -30,11 +34,27 @@ namespace BusinesLogicLayer.Implementation
             fromUser = unit.UserRepository.SearchByUsername(fromUser);
             Post likedPost = new Post { PostId = postId };
             likedPost = unit.PostRepository.SearchById(likedPost);
+            if (fromUser.Id == likedPost.UserId)
+            {
+                return;
+            }
             not.FromWhoId = fromUser.Id;
             not.ForWhoId = likedPost.UserId;
             not.PostId = postId;
             unit.LikeNotificationsRepository.Delete(not);
             unit.Save();
+        }
+
+        public List<LikeNotificationDto> GetAllForUser(int id)
+        {
+            User u = new User { Id = id };
+            List<LikeNotification> notif = unit.LikeNotificationsRepository.GetAllForUser(u).OfType<LikeNotification>().ToList();
+            List<LikeNotificationDto> notifDto = new List<LikeNotificationDto>();
+            notif.ForEach(n =>
+            {
+                notifDto.Add(mapper.toDto(n));
+            });
+            return notifDto;
         }
 
         public void SendLikeNotification(int postId, string username)
@@ -47,6 +67,10 @@ namespace BusinesLogicLayer.Implementation
             not.Date = DateTime.Now;
             Post likedPost = new Post { PostId = postId };
             likedPost = unit.PostRepository.SearchById(likedPost);
+            if(fromUser.Id==likedPost.UserId)
+            {
+                return;
+            }
             not.ForWhoId = likedPost.UserId;
             unit.LikeNotificationsRepository.Add(not);
             unit.Save();
